+++ src/components/Project/Resource/DFR3MappingPage.tsx
@@ -150,11 +150,11 @@ const DFR3MappingPage = (): JSX.Element => {
                         />
                         <ProjectHeader project={project} />
                         <Divider />
-                        <Grid container spacing={5} mt={3} ml={0}>
-                            <Grid sm={3}>
+                        <Grid container spacing={2} mt={3} ml={0}>
+                            <Grid sm={2}>
                                 <ProjectSidebar id={project.id} />
                             </Grid>
-                            <Grid sm={9}>
+                            <Grid sm={10}>
                                 <ResourceFilterBar
                                     title="DFR3Mappings"
                                     icon={<DFR3Icon sx={{ verticalAlign: "middle" }} />}