+++ src/store/hooks.ts
@@ -210,16 +210,14 @@ export const useUserUsageStats = () => {
                 const usageValue = {
                     hazards: {
                         entities: {
-                            text: `${usage.total_number_of_hazards ?? 0}\n/\n${
-                                allocations.total_number_of_hazards ?? 0
-                            }`,
+                            text: `${usage.total_number_of_hazards ?? 0}/${allocations.total_number_of_hazards ?? 0}`,
                             value: getValue(
                                 usage.total_number_of_hazards ?? 0,
                                 allocations.total_number_of_hazards ?? 0
                             )
                         },
                         disk: {
-                            text: `${usage.total_file_size_of_hazard_datasets}\n of\n ${allocations.total_file_size_of_hazard_datasets}`,
+                            text: `${usage.total_file_size_of_hazard_datasets} of ${allocations.total_file_size_of_hazard_datasets}`,
                             value: getValue(
                                 usage.total_file_size_of_hazard_datasets_byte ?? 0,
                                 allocations.total_file_size_of_hazard_datasets_byte ?? 0
@@ -228,16 +226,14 @@ export const useUserUsageStats = () => {
                     },
                     datasets: {
                         entities: {
-                            text: `${usage.total_number_of_datasets ?? 0}\n/\n${
-                                allocations.total_number_of_datasets ?? 0
-                            }`,
+                            text: `${usage.total_number_of_datasets ?? 0}/${allocations.total_number_of_datasets ?? 0}`,
                             value: getValue(
                                 usage.total_number_of_datasets ?? 0,
                                 allocations.total_number_of_datasets ?? 0
                             )
                         },
                         disk: {
-                            text: `${usage.total_file_size_of_datasets}\n of\n ${allocations.total_file_size_of_datasets}`,
+                            text: `${usage.total_file_size_of_datasets} of ${allocations.total_file_size_of_datasets}`,
                             value: getValue(
                                 usage.total_file_size_of_datasets_byte ?? 0,
                                 allocations.total_file_size_of_datasets_byte ?? 0
@@ -246,7 +242,7 @@ export const useUserUsageStats = () => {
                     },
                     dfr3: {
                         entities: {
-                            text: `${usage.total_number_of_dfr3 ?? 0}\n/\n${allocations.total_number_of_dfr3 ?? 0}`,
+                            text: `${usage.total_number_of_dfr3 ?? 0}/${allocations.total_number_of_dfr3 ?? 0}`,
                             value: getValue(usage.total_number_of_dfr3 ?? 0, allocations.total_number_of_dfr3 ?? 0)
                         }
                     }
@@ -352,9 +348,11 @@ export const useHazardStats = (projectHazards: Hazard[]) => {
 export const useWorkflowAndExecutionCount = (projectWorkflows: Workflow[]) => {
     const [workflowCount, setWorkflowCount] = React.useState<number>(0);
     const [executionCount, setExecutionCount] = React.useState<number>(0);
+    const [execByWorkflow, setExecByWorkflow] = React.useState<{ itemName: string; count: number }[]>([]);
 
     React.useEffect(() => {
         setWorkflowCount(projectWorkflows.length);
+        const execByWorkflowTemp: { itemName: string; count: number }[] = [];
         const wfIds = projectWorkflows.map((wf) => wf.id);
         const fetchExecutionCount = async () => {
             try {
@@ -364,11 +362,22 @@ export const useWorkflowAndExecutionCount = (projectWorkflows: Workflow[]) => {
                     })
                 );
                 const responses = await Promise.all(requests);
-                let execCount = responses.reduce((acc, response) => {
+                const execCount = responses.reduce((acc, response) => {
                     const dataLength = Array.isArray(response.data) ? response.data.length : 0;
                     return acc + dataLength;
                 }, 0);
                 setExecutionCount(execCount);
+                responses.forEach((response, i) => {
+                    if (!Array.isArray(response.data) || response.data.length === 0) {
+                        execByWorkflowTemp.push({ itemName: projectWorkflows[i]?.title || "Unknown", count: 0 });
+                    }
+                    const wfId = response.data[0]?.workflowId;
+                    const workflow = projectWorkflows.find((wf) => wf.id === wfId);
+                    if (workflow) {
+                        execByWorkflowTemp.push({ itemName: workflow.title, count: response.data.length });
+                    }
+                });
+                setExecByWorkflow(execByWorkflowTemp);
             } catch (error) {
                 console.error("Error fetching execution count: ", error);
             }
@@ -376,7 +385,7 @@ export const useWorkflowAndExecutionCount = (projectWorkflows: Workflow[]) => {
         fetchExecutionCount();
     }, [projectWorkflows]);
 
-    return { workflowCount, executionCount };
+    return { workflowCount, executionCount, execByWorkflow };
 };
 
 export const useWorkflowAutoSave = (workflow: DatawolfWorkflowFile, workflowID: string | null, interval: number) => {